+++ R/type_convert.R
@@ -6,6 +6,7 @@
 #'
 #' @param df A data frame.
 #' @inheritParams col_types_standardise
+#' @inheritParams tokenizer_delim
 #' @export
 #' @examples
 #' df <- data.frame(
@@ -16,14 +17,22 @@
 #' str(df)
 #' str(type_convert(df))
 #'
-#' df <- data.frame(x = c("NA", "10"), stringsAsFactors=  FALSE)
-#' type_convert(df)
-type_convert <- function(df, col_types = NULL) {
+#' df <- data.frame(x = c("NA", "10"), stringsAsFactors = FALSE)
+#' str(type_convert(df))
+type_convert <- function(df, col_types = NULL, na = c("", "NA"), trim_ws = TRUE) {
   is_character <- vapply(df, is.character, logical(1))
 
-  guesses <- lapply(df[is_character], collectorGuess)
-  col_types <- col_types_standardise(col_types, names(df), guesses)
+  char_cols <- df[is_character]
+  guesses <- lapply(char_cols, function(x) {
+    x[x %in% na] <- NA
+    collectorGuess(x)
+  })
+  col_types <- col_types_standardise(col_types, names(char_cols), guesses)
 
-  df[is_character] <- Map(type_convert_col, df[is_character], col_types, which(is_character))
+  df[is_character] <- lapply(seq_along(char_cols), function(i) {
+    type_convert_col(char_cols[[i]], col_types[[i]], which(is_character)[i],
+      na = na, trim_ws = trim_ws)
+  })
+
   df
 }